Adicionar o objeto nos argumentos, colocar no classpath ao iniciar a Lori
