+++ src/main/java/com/mrpowergamerbr/loritta/LorittaLauncher.kt
@@ -5,6 +5,9 @@ import com.mrpowergamerbr.loritta.utils.WebsiteUtils
 import com.mrpowergamerbr.loritta.utils.config.LorittaConfig
 import java.io.File
 import java.io.IOException
+import java.nio.file.Paths
+import java.util.jar.Attributes
+import java.util.jar.JarFile
 
 /**
  * Loritta's Launcher
@@ -28,6 +31,23 @@ object LorittaLauncher {
 		// https://bugs.openjdk.java.net/browse/JDK-7016595
 		// Nós precisamos ativar o PATCH manualmente
 		WebsiteUtils.allowMethods("PATCH")
+
+		// https://www.reddit.com/r/Kotlin/comments/8qdd4x/kotlin_script_engine_and_your_classpaths_what/
+		val path = this::class.java.protectionDomain.codeSource.location.path
+		val jar = JarFile(path)
+		val mf = jar.manifest
+		val mattr = mf.mainAttributes
+		// Yes, you SHOULD USE Attributes.Name.CLASS_PATH! Don't try using "Class-Path", it won't work!
+		val manifestClassPath = mattr[Attributes.Name.CLASS_PATH] as String
+
+		// The format within the Class-Path attribute is different than the one expected by the property, so let's fix it!
+		// By the way, don't forget to append your original JAR at the end of the string!
+		val clazz = LorittaLauncher::class.java
+		val protectionDomain = clazz.protectionDomain
+		val propClassPath = manifestClassPath.replace(" ", ":") + ":${Paths.get(protectionDomain.codeSource.location.toURI()).fileName}"
+
+		// Now we set it to our own classpath
+		System.setProperty("kotlin.script.classpath", propClassPath)
 
 		val file = File(System.getProperty("conf") ?: "./config.yml")
 		var config: LorittaConfig? = null
